pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tygent"
version = "0.4.0"
description = "Transform LLM Agents into High-Performance Engines with DAG optimization"
readme = "README.md"
authors = [
    {name = "Tygent AI", email = "info@tygent.ai"},
]
license = {text = "CC BY-NC 4.0"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "aiohttp>=3.8",
    "python-dotenv",
    "google-adk; python_version >= '3.9'",
    "google-genai; python_version >= '3.9'",
    "pytest>=8.3.5",
    "black>=24.8.0",
    "isort>=5.13.2",
]

[project.urls]
"Homepage" = "https://github.com/tygent-ai/tygent-py"
"Bug Tracker" = "https://github.com/tygent-ai/tygent-py/issues"
"Documentation" = "https://tygent.ai/docs"

[tool.pytest]
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]

[tool.isort]
profile = "black"
